requirements.txt
Flask==3.0.0
selenium==4.15.2
webdriver-manager==4.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Cors==5.0.0
pymongo==4.10.1